[Tutorial] Extract createReplyMessages and createFeedBackMessage functions from choosingReply
