+++ src/webhook/handlers/choosingReply.js
@@ -1,17 +1,82 @@
 import { t } from 'ttag';
 import gql from 'src/lib/gql';
-import { getArticleURL } from 'src/lib/sharedUtils';
 import {
-  createReferenceWords,
-  createTypeWords,
-  ellipsis,
   getLIFFURL,
   ManipulationError,
   FLEX_MESSAGE_ALT_TEXT,
   createNotificationSettingsBubble,
+  createReplyMessages,
 } from './utils';
 import ga from 'src/lib/ga';
 import UserSettings from 'src/database/models/userSettings';
+
+/**
+ * @param {*} userId LINE user ID that does the input
+ * @returns {object} Flex message object
+ */
+async function createAskReplyFeedbackMessage(userId) {
+  const helpfulTitle = t`Is the reply helpful?`;
+  const { allowNewReplyUpdate } = await UserSettings.findOrInsertByUserId(
+    userId
+  );
+
+  return {
+    type: 'flex',
+    altText: helpfulTitle + '\n' + FLEX_MESSAGE_ALT_TEXT,
+    contents: {
+      type: 'carousel',
+      contents: [
+        {
+          type: 'bubble',
+          body: {
+            type: 'box',
+            layout: 'vertical',
+            paddingAll: 'xl',
+            contents: [
+              {
+                type: 'text',
+                wrap: true,
+                text: helpfulTitle,
+              },
+            ],
+          },
+          footer: {
+            type: 'box',
+            layout: 'horizontal',
+            spacing: 'sm',
+            contents: [
+              {
+                type: 'button',
+                style: 'primary',
+                color: '#ffb600',
+                action: {
+                  type: 'uri',
+                  label: t`Yes`,
+                  uri: getLIFFURL('feedback/yes', userId, data.sessionId),
+                },
+              },
+              {
+                type: 'button',
+                style: 'primary',
+                color: '#ffb600',
+                action: {
+                  type: 'uri',
+                  label: t`No`,
+                  uri: getLIFFURL('feedback/no', userId, data.sessionId),
+                },
+              },
+            ],
+          },
+        },
+        // Ask user to turn on notification if the user did not turn it on
+        //
+        process.env.NOTIFY_METHOD &&
+          !allowNewReplyUpdate &&
+          createNotificationSettingsBubble(),
+      ].filter(m => m),
+    },
+  };
+}
 
 export default async function choosingReply(params) {
   let { data, state, event, issuedAt, userId, replies, isSkipUser } = params;
@@ -44,93 +109,11 @@ export default async function choosingReply(params) {
 
   const { GetReply, GetArticle } = getReplyData;
 
-  const articleUrl = getArticleURL(data.selectedArticleId);
-  const typeStr = createTypeWords(GetReply.type).toLocaleLowerCase();
-  const helpfulTitle = t`Is the reply helpful?`;
-  const { allowNewReplyUpdate } = await UserSettings.findOrInsertByUserId(
-    userId
-  );
-
-  replies = [
-    {
-      type: 'text',
-      text: `💡 ${t`Someone on the internet replies to the message:`}`,
-    },
-    {
-      type: 'text',
-      text: ellipsis(GetReply.text, 2000),
-    },
-    {
-      type: 'text',
-      text: ellipsis(createReferenceWords(GetReply), 2000),
-    },
-    {
-      type: 'text',
-      text:
-        `⬆️ ${t`Therefore, the author think the message ${typeStr}.`}\n\n` +
-        `💁 ${t`These messages are provided by some nice volunteers. Please refer to their references and make judgements on your own.`}\n` +
-        (GetArticle.replyCount > 1
-          ? `🗣️ ${t`There are different replies for the message. Read them all here before making judgements:`}\n${articleUrl}\n`
-          : '') +
-        `\n⁉️ ${t`If you have different thoughts, you may have your say here:`}\n${articleUrl}`,
-    },
-    {
-      type: 'flex',
-      altText: helpfulTitle + '\n' + FLEX_MESSAGE_ALT_TEXT,
-      contents: {
-        type: 'carousel',
-        contents: [
-          {
-            type: 'bubble',
-            body: {
-              type: 'box',
-              layout: 'vertical',
-              paddingAll: 'xl',
-              contents: [
-                {
-                  type: 'text',
-                  wrap: true,
-                  text: helpfulTitle,
-                },
-              ],
-            },
-            footer: {
-              type: 'box',
-              layout: 'horizontal',
-              spacing: 'sm',
-              contents: [
-                {
-                  type: 'button',
-                  style: 'primary',
-                  color: '#ffb600',
-                  action: {
-                    type: 'uri',
-                    label: t`Yes`,
-                    uri: getLIFFURL('feedback/yes', userId, data.sessionId),
-                  },
-                },
-                {
-                  type: 'button',
-                  style: 'primary',
-                  color: '#ffb600',
-                  action: {
-                    type: 'uri',
-                    label: t`No`,
-                    uri: getLIFFURL('feedback/no', userId, data.sessionId),
-                  },
-                },
-              ],
-            },
-          },
-          // Ask user to turn on notification if the user did not turn it on
-          //
-          process.env.NOTIFY_METHOD &&
-            !allowNewReplyUpdate &&
-            createNotificationSettingsBubble(),
-        ].filter(m => m),
-      },
-    },
-  ];
+  replies = createReplyMessages(
+    GetReply,
+    GetArticle,
+    data.selectedArticleId
+  ).concat(await createAskReplyFeedbackMessage(userId));
 
   const visitor = ga(userId, state, data.selectedArticleText);
   // Track when user select a reply.
